requirements.txt
Flask
mysql-connector-python
pandas
python-dotenv
SQLAlchemy